requirements.txt
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
lxml==5.4.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
peewee==3.15.4
platformdirs==4.3.8
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
soupsieve==2.7
telebot==0.0.5
typing_extensions==4.13.2
urllib3==2.4.0
